feat(leetcode-daily): 2024-01-05

+++ leetcode-rs/examples/1696. 跳跃游戏 VI.rs
@@ -7,9 +7,9 @@ impl Solution {
         let mut dp = vec![0; n];
         dp[0] = nums[0];
         let mut q = std::collections::VecDeque::new();
-        q.push_back(0usize);
+        q.push_back(0);
         for i in 1..n {
-            if i - q.front().unwrap() > k as usize {
+            if i - q.front().unwrap() > k {
                 q.pop_front();
             }
             dp[i] = dp[q.front().unwrap()] + nums[i];
